Fixed vertical alignment issue by setting the image size.
Added an information button to the swiping_page.dart so we can view more personal details about the dog profile we are viewing.

Updated main so that it contains new information about each profile.

Added a temporary placeholder for where the heart_counter.dart heart icon will navigate to in matches.dart as a new widget.

Signed-off-by: onalanb <baran.onalan@gmail.com>